fix(add-item-flow): function move separate file
